Make CSV timestampFormat selector a component for other transformers
